+++ docs/overview.rst
@@ -88,8 +88,45 @@ programming easier & accessible:
    energy, you can easily get the forces on each elements.
    See :ref:`differentiable` for more details.
 
+
+Summary
+-------
+
+Example
+*******
+
+Here we showcase an example on how to render a classical image via Taichi:
+
+.. code-block:: python
+
+    import taichi as ti
+
+    # declare a 512x512 field whose elements are 3D vectors (RGB channels)
+    rgb_image = ti.Vector.field(3, dtype=ti.f32, shape=(512, 512))
+
+
+    @ti.kernel  # functions decorated by @ti.kernel will be compiled by Taichi
+    def render():
+        for i, j in rgb_image:  # iterate over each pixels in the 512x512 field
+            r = i / 512
+            g = j / 512
+            b = 0.0
+            # set the vector value at [i, j] in the field:
+            rgb_image[i, j] = ti.Vector([r, g, b])
+
+
+    gui = ti.GUI('UV', (512, 512))  # create a 512x512 window
+    while gui.running:
+        render()
+        gui.set_image(rgb_image)  # display the rendered image
+        gui.show()
+
+
+See :ref:`install` for more details about how to install Taichi via ``pip``.
+See :ref:`hello` for more details about Taichi langurage and syntax.
+
 Features
---------
+********
 
 To sum up, Taichi provides you with:
 
@@ -102,7 +139,7 @@ To sum up, Taichi provides you with:
 - Differentiable programming
 
 Design decisions
-----------------
+****************
 
 - Decouple computation from data structures
 - Domain-specific compiler optimizations
